requirements.txt
asyncio==3.4.3
six==1.15.0
websocket-client==0.57.0
